requirements.txt
certifi==2018.11.29
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
Django==2.1.5
django-filter==2.1.0
djangorestframework==3.9.1
httpie==1.0.2
idna==2.8
itypes==1.1.0
Jinja2==2.10
logilab-common==1.4.2
logilab-database==1.15.0
Markdown==3.0.1
MarkupSafe==1.1.0
Pygments==2.3.1
python-dateutil==2.8.0
pytz==2018.9
PyYAML==3.13
requests==2.21.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
yams==0.45.1
Yapps2==2.2.0
